회원가입

+++ src/main/java/com/ll/exam/sbb/answer/AnswerController.java
@@ -24,7 +24,7 @@ public class AnswerController {
     private final AnswerService answerService;
 
     @PostMapping("/create/{id}")
-    public String createAnswer(Model model, @PathVariable int id, @Valid AnswerForm answerForm , BindingResult bindingResult) {
+    public String createAnswer(Model model, @PathVariable Long id, @Valid AnswerForm answerForm , BindingResult bindingResult) {
 
         Question question = questionService.getQuestion(id);
